+++ tests/testthat/test-MCMCEnsemble.R
@@ -3,7 +3,7 @@ p.log <- function(x) {
   return(-x[1]^2/200 - 1/2*(x[2]+B*x[1]^2-100*B)^2)
 }
 
-test_that("mcmcensemble", {
+test_that("output sizes and names", {
 
   res1 <- MCMCEnsemble(p.log, lower.inits=c(a=0, b=0), upper.inits=c(a=1, b=1),
                        max.iter=3000, n.walkers=10, method="s")
@@ -14,18 +14,25 @@ test_that("mcmcensemble", {
 
   expect_identical(dim(res1$samples), c(10L, 300L, 2L))
   expect_identical(dim(res1$samples)[1:2], dim(res1$log.p))
-
   expect_identical(dimnames(res1$samples)[[3]], c("a", "b"))
+  expect_identical(attr(res1, "ensemble.sampler"), "stretch")
 
   res2 <- MCMCEnsemble(p.log, lower.inits=c(0, 0), upper.inits=c(a=1, b=1),
                        max.iter=3000, n.walkers=10, method="s")
 
   expect_identical(dimnames(res2$samples)[[3]], c("para_1", "para_2"))
 
+})
+
+test_that("coda", {
+
+  skip_if_not_installed("coda")
+
   res3 <- MCMCEnsemble(p.log, lower.inits=c(a=0, b=0), upper.inits=c(a=1, b=1),
                        max.iter=3000, n.walkers=10, method="d", coda=TRUE)
 
   expect_s3_class(res3$samples, "mcmc.list")
+  expect_identical(attr(res3, "ensemble.sampler"), "differential.evolution")
 
 })
 
@@ -59,4 +66,24 @@ test_that("errors", {
                  max.iter=3000, n.walkers=10, method="d"),
     "numeric of length 1"
   )
+})
+
+test_that("named arguments", {
+
+  set.seed(20200111)
+  res1 <- MCMCEnsemble(p.log, lower.inits=c(a=0, b=0), upper.inits=c(a=1, b=1),
+                       max.iter=3000, n.walkers=10, method="s")
+
+  p.log.named <- function(x) {
+    B <- 0.03
+    return(-x["a"]^2/200 - 1/2*(x["b"]+B*x["a"]^2-100*B)^2)
+  }
+
+  set.seed(20200111)
+  res2 <- MCMCEnsemble(p.log.named, lower.inits=c(a=0, b=0),
+                       upper.inits=c(a=1, b=1), max.iter=3000, n.walkers=10,
+                       method="s")
+
+  expect_identical(res1, res2)
+
 })